Improved startup library
Improved str_from(): Added parameter "required" which requires the needle to exist for the function to return anything at all
Improved page_show(): Now tests if the requested page exists. If not, a 404 exception will be thrown. If the 404 page does not exist, a general exception will be thrown instead