(WIP) Overhauling update script
The update script is being completely rewritten in both content and functionality. The new update script will auto find the available phoundation project, compare versions of the current project with the phoundation project, and only update if updates are available within the same version lines. So if, for example, the project is on 2.2.4, and on phoundation, the versions 2.2.5, 2.2.6, and 2.3.0 are available, the update script by default will update to 2.2.6, unless --upgrade is specified. There will be way more options available, but this will be the first aim. Once this is functional, projects will always only update within their own version line, which should cause way less compatibility issues