Improved Core object
php error handling will now no longer cause state change and notifications. This will instead be left to the uncaught exception handler. This way the system will be able to catch PHP errors that are not directly fatal without the Core being left in an invalid state and having nonsense notifications sent out