Improved functions library file
Added get_numeric() Returns either a float number, int number, or NULL, depending on given source. If the given source can be interpreted as an integer, that is what will be returned, same for float. If either fail, NULL will be returned

Fixed execute() will now only echo output if the output is a string or RenderInterface object

Cleaned code

Signed-off-by: Sven Olaf Oostenbrink <so.oostenbrink@gmail.com>