Fixed weird phpstorm issues
So phpstorm doesnt really understand how SOLID works and keeps complaining about incompatibilities, even though in reality all is fine and PHP itself has no problem either.

In any case, it causes a lot of false errors, so I fixed the datatypes for the Data objects to comply with storms nagging.

phpstorm is just the worst.

Signed-off-by: Sven Olaf Oostenbrink <sven@medinet.ca>