Improved Data library
Added optional parameter "auto_save" to setStatus, and to all methods that primarily call getStatus, such as delete and undelete. This change was made for DataEntries and DataIterators. New property of DataEntries called "modified_entries" which will store the number of entries that were modified, to replace the return value of delete()

Signed-off-by: Harrison Macey <harrison@medinet.ca>